[UTP AGP] added workaround for dslName
So added the deviceName after the device is set in the device
controller. This will guarantee that the device properties are not
overwritten by the controller.

This is a stop gap in the interim, until we add a clearer way of
communicating the metadata from one plugin to another. I've made a bug
to follow up on this issue.

Bug: 173824087
Test: verified the device-info.pb contains the dslName.
Change-Id: I5b5a53cd47129292dabca4ff5820dd5a6635113a