[lint] Improve detection of this-expressions
The semantics of VariableExpression.THIS_EXPRESSION have, in practice,
changed between Groovy 2 and Groovy 3: whereas previously it was used
for all `this` expressions, implicitly or not, it is now used only
internally and distinct (in the sense of object identity)
VariableExpressions used for distinct expressions.  This is now
documented (in a comment) as being required to attach source information
to each such expression, and so we can expect this to continue being the
case.

The good news is that there is a documented way to check for a
VariableExpression being associated with `this`, so use that instead.
(It does in fact currently boil down to a String check, but using
Groovy's string check hidden behind a method rather than ours might make
this less likely to break when Groovy 4 is added to the mix...)

Fixes: 183971877
Test: existing, no change in behaviour
Change-Id: I985ac510dc1b82993695ec3544d94f4499535669