[AGP UTP] added backup way to close emulator
Added a way to close the emulator outside of closing the emulator
process. This was necessary as sometimes on Windows closing the process
that spawned the emulator may not necessarily close the emulator itself
leading to issues.

After attempting to close the emulator by closing the strarting process,
we search avd for the serial on avd (assuming a failed close). If the
emulator still exists on ADB we attempt to kill the process on ADB via
the following command:

adb -s <serial-number> emu kill

Bug: 183246439
Test: added tests to the launcher test to verify exact behavior. Plus
did some basic testing on my windows computer.

Change-Id: Iabbb93186e6534bfd9c52fe19f7dcdf39b7db33c