Make ninja verbosity respect CMAKE_VERBOSE_MAKEFILE=1
We bypass this flag by calling ninja directly rather than calling
"CMake build ...". This CL detects the case that user has set
CMAKE_VERBOSE_MAKEFILE property in build.gradle arguments and,
when the value is CMake-truthy, it sets the -v flag on the ninja
command-line.

Note that we can't control this behavior for fork CMake, so there
is no integration test for it. However, fork CMake does call
"CMake build" and it does respected CMAKE_VERBOSE_MAKEFILE because
of that.

Bug: 184060944
Test: Wrote an integration test

Change-Id: Ied92b6050ece3a0b6a2698d70bf77e3ed01e1df7