Add maven coordinate info to local AAR for lint
Previously, some lint model files listed the groupId of the resolved
maven coordinates of module dependencies as "artifacts", which resulted
in false positives from the RestrictToDetector.

This change adds the maven coordinate info to the local AAR for lint,
which allows the correct groupId to be specified in the lint models.

Fixes: 180408990
Test: LintRestrictedApiTest, LintModelMetadataTaskTest
Change-Id: If9c8dd94b36c5c07eebb358b5bd5a5cc40d19e9d