Use config=rcache for coverage pre/post-processing
Coverage runs are slower end-to-end in practice, but not their test
phases and not when run while RBE is empty. In practice, RBE is usually
full so we suspect that the post-processing invocation enters at the
back of the line and then has to wait before executing and that this
wait is the cause of the slower end-to-end behavior.

By using config=rcache, post-processing work should take place locally
and avoid this issue. Post-processing is hermetic so this shouldn't
matter, but previously we had some mysterious issues when running
locally (b/177232156). However, that used local execution whereas this
uses linux-sandbox so it may work better. We should still keep any eye
out in case of trouble.

The short pre-processing step was converted as well as there's no real
reason it needs to be remote.

Test: N/A
Bug: 186584583
Change-Id: Ia9005cf51c08b1d76bc4cb63416ec97356d7a70b