Fix prefab bug introduced in refactoring.
`list != null` is not the same as `!list.isEmpty()`.

I can't fully restore the earlier behavior because this function is
now also called during the configuration stage to decide whether or
not to pass CMAKE_FIND_ROOT_PATH or not, and we don't know whether or
not prefab will be called until after configurations are resolved.
I've tried to preserve some of the old behavior by only passing that
value when the prefab build feature is enabled and we're using a new
enough NDK, but it will still pass the property even if there are no
dependencies to import.

Test: Added regression test
Bug: https://issuetracker.google.com/183634734
Change-Id: I5883d90498afe073154b24c3529a6472d05668d5