Enforce the ownership of gRPC socket's fd
As the gRPC library would make a copy of the fd,
the agent should tak the full control of the fd's
lifetime and close() it when needed.

Logs the close() call if there's an error.

Also makes the DaemonStatusChanged callback configurable
because sometimes it should be thrown away after execution
but sometimes it shouldn't.

Bug: 186054748
Test: manual; used logs to verify the fix
Change-Id: I92611bb09cbbe01ef8500da2618ffd2bca2f47d0