188464656: Quickfixes do not work with Kotlin's property syntax
The original bug here was that with a SourceLockedOrientationActivity
warning from Kotlin code, the SARIF report was truncated. This CL fixes
that by precomputing the fixes and only keeping them in the report if
they are correctly computed (there is some runtime checking of fixes and
that's where it can discover conflicts.)

However, the deeper issue is that in many cases, if you just use the
default UAST range for a call, you'll only pick up the left hand side of
a method call when using the property syntax. However, the argument list
*does* contain elements which point to the right hand side.  Therefore,
this CL updates the range location code for looking up call ranges to
(when includeArguments is true) merge the left and right hand side
ranges for the last argument, call node, and optionally the receiver
node.

Fixes: 188464656
Test: New unit test (plus existing to verify location code still works)
Change-Id: Iaf2fbc840426a84c6ddbff970cde3959338247df