Clarify interoperability detector and improve HTML rendering
The interoperability detector for mismatched return types had both a
misleading error message (mentioning the "setter" instead of the super
getter) and the linked location was for the (unrelated to this issue)
setter, instead of the super getter.

Furthermore, the HTML report rendering of linked locations did not look
good; the end offset of the errors were not used, so it would always be
a line trailing out to the end of the code box. This CL also fixes that
issue and updates the tests.

Test: Unit tests included
Bug: Not filed
Change-Id: I84e442172ab1eea04a3e071683d989364c683921